Break circular dependency for `parseRule`
